Added more javadoc
