+++ .idea/misc.xml
@@ -4,17 +4,6 @@
     <option name="filePathToZoomLevelMap">
       <map>
         <entry key="..\:/1Projects/Walter/School/2021-2022/Winter/cmput301/team_project/QRGo/app/src/main/res/layout/activity_main.xml" value="0.22826086956521738" />
-        <entry key="..\:/Users/moham/AndroidStudioProjects/QRGo/app/src/main/res/drawable/ic_outline_photo_camera_24.xml" value="0.1035" />
-        <entry key="..\:/Users/moham/AndroidStudioProjects/QRGo/app/src/main/res/layout/activity_main.xml" value="0.22" />
-        <entry key="..\:/Users/Amro Amanuddein/AndroidStudioProjects/QRGo/app/src/main/res/drawable/rounded_corners.xml" value="0.2365" />
-        <entry key="..\:/Users/Amro Amanuddein/AndroidStudioProjects/QRGo/app/src/main/res/layout/activity_login.xml" value="0.264" />
-        <entry key="..\:/Users/Amro Amanuddein/AndroidStudioProjects/QRGo/app/src/main/res/layout/activity_login_scan.xml" value="0.1" />
-        <entry key="..\:/Users/moham/AndroidStudioProjects/QRGo/app/src/main/res/drawable/ic_outline_photo_camera_24.xml" value="0.108" />
-        <entry key="..\:/Users/moham/AndroidStudioProjects/QRGo/app/src/main/res/layout/activity_main.xml" value="0.3231578947368421" />
-        <entry key="..\:/Users/moham/AndroidStudioProjects/QRGo/app/src/main/res/layout/fragment_camera.xml" value="0.3333333333333333" />
-        <entry key="..\:/Users/moham/AndroidStudioProjects/QRGo/app/src/main/res/layout/fragment_profile.xml" value="0.19083969465648856" />
-        <entry key="..\:/Users/moham/AndroidStudioProjects/QRGo/app/src/main/res/layout/profile_qr_code_list_item.xml" value="0.4466666666666667" />
-        <entry key="..\:/Users/moham/AndroidStudioProjects/QRGo/app/src/main/res/menu/bottom_nav_menu.xml" value="0.4466666666666667" />
         <entry key="..\:/Users/ervin/CMPUT301/QRGo/app/src/main/res/layout/activity_main.xml" value="0.26875" />
         <entry key="..\:/Users/ervin/CMPUT301/QRGo/app/src/main/res/layout/fragment_camera.xml" value="0.26875" />
         <entry key="..\:/Users/ervin/CMPUT301/QRGo/app/src/main/res/layout/fragment_leaderboard.xml" value="0.26875" />
@@ -31,24 +20,14 @@
         <entry key="..\:/Users/smol/Documents/school/2021-2022/Winter/CMPUT_301/team_project/QRGo/app/src/main/res/drawable/ic_home_black_24dp.xml" value="0.252" />
         <entry key="..\:/Users/smol/Documents/school/2021-2022/Winter/CMPUT_301/team_project/QRGo/app/src/main/res/drawable/ic_launcher_background.xml" value="0.252" />
         <entry key="..\:/Users/smol/Documents/school/2021-2022/Winter/CMPUT_301/team_project/QRGo/app/src/main/res/drawable/ic_notifications_black_24dp.xml" value="0.252" />
-        <entry key="..\:/Users/smol/Documents/school/2021-2022/Winter/CMPUT_301/team_project/QRGo/app/src/main/res/layout/activity_main.xml" value="0.48020833333333335" />
         <entry key="..\:/Users/smol/Documents/school/2021-2022/Winter/CMPUT_301/team_project/QRGo/app/src/main/res/drawable/ic_outline_photo_camera_24.xml" value="0.114" />
         <entry key="..\:/Users/smol/Documents/school/2021-2022/Winter/CMPUT_301/team_project/QRGo/app/src/main/res/drawable/ic_profile_24.xml" value="0.114" />
         <entry key="..\:/Users/smol/Documents/school/2021-2022/Winter/CMPUT_301/team_project/QRGo/app/src/main/res/drawable/ic_qr_code_scanner_24.xml" value="0.114" />
-        <entry key="..\:/Users/smol/Documents/school/2021-2022/Winter/CMPUT_301/team_project/QRGo/app/src/main/res/layout/activity_main.xml" value="0.1" />
+        <entry key="..\:/Users/smol/Documents/school/2021-2022/Winter/CMPUT_301/team_project/QRGo/app/src/main/res/layout/activity_main.xml" value="0.4" />
         <entry key="..\:/Users/smol/Documents/school/2021-2022/Winter/CMPUT_301/team_project/QRGo/app/src/main/res/layout/fragment_dashboard.xml" value="0.10462962962962963" />
         <entry key="..\:/Users/smol/Documents/school/2021-2022/Winter/CMPUT_301/team_project/QRGo/app/src/main/res/layout/fragment_leaderboard.xml" value="0.28703703703703703" />
         <entry key="..\:/Users/smol/Documents/school/2021-2022/Winter/CMPUT_301/team_project/QRGo/app/src/main/res/layout/fragment_map.xml" value="0.1" />
         <entry key="..\:/Users/smol/Documents/school/2021-2022/Winter/CMPUT_301/team_project/QRGo/app/src/main/res/layout/fragment_notifications.xml" value="0.23333333333333334" />
-        <entry key="..\:/Users/smol/Documents/school/2021-2022/Winter/CMPUT_301/team_project/QRGo/app/src/main/res/layout/fragment_profile.xml" value="0.28703703703703703" />
-        <entry key="..\:/Users/smol/Documents/school/2021-2022/Winter/CMPUT_301/team_project/QRGo/app/src/main/res/layout/fragment_search.xml" value="0.1" />
-        <entry key="..\:/Users/smol/Documents/school/2021-2022/Winter/CMPUT_301/team_project/QRGo/app/src/main/res/menu/bottom_nav_menu.xml" value="0.23333333333333334" />
-        <entry key="app/src/main/res/layout/activity_main.xml" value="0.375" />
-        <entry key="app/src/main/res/layout/fragment_camera.xml" value="0.4979166666666667" />
-        <entry key="app/src/main/res/layout/fragment_leaderboard.xml" value="0.32708333333333334" />
-        <entry key="app/src/main/res/layout/fragment_map.xml" value="0.36018518518518516" />
-        <entry key="app/src/main/res/layout/fragment_profile.xml" value="0.4979166666666667" />
-        <entry key="app/src/main/res/layout/leaderboard_list_item.xml" value="0.32708333333333334" />
         <entry key="..\:/Users/smol/Documents/school/2021-2022/Winter/CMPUT_301/team_project/QRGo/app/src/main/res/layout/fragment_profile.xml" value="0.4" />
         <entry key="..\:/Users/smol/Documents/school/2021-2022/Winter/CMPUT_301/team_project/QRGo/app/src/main/res/layout/fragment_search.xml" value="0.1" />
         <entry key="..\:/Users/smol/Documents/school/2021-2022/Winter/CMPUT_301/team_project/QRGo/app/src/main/res/layout/leaderboard_list_item.xml" value="0.27685185185185185" />
