+++ app/src/main/java/com/example/myapplication/dataClasses/user/Player.java
@@ -10,8 +10,8 @@ import java.util.ArrayList;
  * This class is a representation of the player. It holds information necessary such as type of user, their score,
  * if they are a player or an admin.
  * @author Walter
+ * @author Sankalp
  * @see ScoringQRCode
- * @version 1.0.0
  */
 public class Player {
     private String username;
@@ -21,7 +21,7 @@ public class Player {
     private final boolean isAdmin;
 
     /**
-     * constructor of the player class
+     * Constructor of the player class.
      * @param username the username of the player
      * @param isAdmin whether or not the player has admin privilages
      */
@@ -32,7 +32,7 @@ public class Player {
     }
 
     /**
-     * Clears the qrCode list so that you can add a new list of qr codes
+     * Clears the qrCode list so that you can add a new list of qr codes.
      */
     public void resetQrCodeList() {
         scannedQRCodes.clear();
@@ -112,7 +112,7 @@ public class Player {
 
     /**
      * Allows someone to get the list of qr codes the player has.
-     * @return the list of qr codes on the player profile.
+     * @return The list of qr codes on the player profile.
      */
     public ArrayList<ScoringQRCode> getQrCodes() {
         return this.scannedQRCodes;
